requirements.txt
asyncua==1.0.1
pyyaml==6.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
